+++ src/internal_modules/roc_rtp/headers.h
@@ -116,7 +116,7 @@ public:
     //! Get header size in bytes.
     uint32_t header_size() const {
         roc_panic_if(sizeof(*this) != 12);
-        return sizeof(*this) + num_csrc() * sizeof(uint32_t);
+        return (uint32_t)sizeof(*this) + num_csrc() * (uint32_t)sizeof(uint32_t);
     }
 
     //! Clear header.
@@ -132,8 +132,8 @@ public:
     //! Set version.
     void set_version(Version v) {
         roc_panic_if((v & Flag_VersionMask) != v);
-        flags_ &= ~(Flag_VersionMask << Flag_VersionShift);
-        flags_ |= (v << Flag_VersionShift);
+        flags_ &= (uint8_t) ~(Flag_VersionMask << Flag_VersionShift);
+        flags_ |= ((uint8_t)v << Flag_VersionShift);
     }
 
     //! Get padding flag.
@@ -143,7 +143,7 @@ public:
 
     //! Set padding flag.
     void set_padding(bool v) {
-        flags_ &= ~(Flag_PaddingMask << Flag_PaddingShift);
+        flags_ &= (uint8_t) ~(Flag_PaddingMask << Flag_PaddingShift);
         flags_ |= ((v ? 1 : 0) << Flag_PaddingShift);
     }
 
@@ -165,7 +165,7 @@ public:
     //! Set payload type.
     void set_payload_type(uint8_t pt) {
         roc_panic_if((pt & MPT_PayloadTypeMask) != pt);
-        mpt_ &= ~(MPT_PayloadTypeMask << MPT_PayloadTypeShift);
+        mpt_ &= (uint8_t) ~(MPT_PayloadTypeMask << MPT_PayloadTypeShift);
         mpt_ |= (pt << MPT_PayloadTypeShift);
     }
 
@@ -176,7 +176,7 @@ public:
 
     //! Set marker bit.
     void set_marker(bool m) {
-        mpt_ &= ~(MPT_MarkerMask << MPT_MarkerShift);
+        mpt_ &= (uint8_t) ~(MPT_MarkerMask << MPT_MarkerShift);
         mpt_ |= ((!!m) << MPT_MarkerShift);
     }
